Actualizacion Get Proveedores

+++ J3AMS/J3AMS.Negocio/CategoriaIvaNegocio.cs
@@ -28,7 +28,7 @@ namespace J3AMS.Negocio
                 {
                     listIva.Add(new CategoriaIva()
                     {
-                        Id = (int)_datos.Lector["Id"],
+                        Id = (byte)_datos.Lector["Id"],
 
                         Descripcion = _datos.Lector["Descripcion"] as string ?? string.Empty,
                     });
